requirements.txt
python-telegram-bot==21.9
APScheduler==3.10.4
requests==2.32.3
pytz==2024.1
beautifulsoup4==4.12.3
